Testing Removing idea folder and changing how gui launches
