+++ HealthTrackerApp/.gitignore
@@ -1,5 +1,5 @@
 # Intellij
-.idea/
+.idea
 *.iml
 *.iws
